Boost Unit Testing w/ Clases & CMake
I think its fully set-up to add unit tests. There is a slight
poor point in the CMakeLists.txt. I create a library and an executable
from the exact same files. I tried to get around this but it didn't
work very well. We really only need the library for tests.