Reorganized files and build files
Removed unused Boost libraries. We'll add them as we need to.
The src folder contains files used to build the library. The tests
have their own folder. GenexCLI (previously OnlineSession) is the
command-line interface and does not belong to the core of genex,
that's why it is moved outside. Code for testing in CMake is put in
the root CMakeLists.txt instead of in the test folder so that we can
call ctest at root level.